Fix imports for stable
